refactor: improve typings and cleanup callbacks
