+++ src/components/layout/ErrorButton.tsx
@@ -2,18 +2,22 @@ import type { ErrorButtonState } from '@/types/types';
 import React from 'react';
 
 class ErrorButton extends React.Component<object, ErrorButtonState> {
-  state = {
+  state: ErrorButtonState = {
     shouldThrow: false
   };
 
-  render(): React.ReactNode {
+  handleClick = (): void => {
+    this.setState({ shouldThrow: true });
+  };
+
+  render(): React.JSX.Element {
     if (this.state.shouldThrow) {
       throw new Error('ErrorBoundary check');
     }
 
     return (
       <button
-        onClick={() => this.setState({ shouldThrow: true })}
+        onClick={this.handleClick}
         className="inline-flex items-center animate-pulse bg-amber-300 text-black border-0 py-1.5 px-3 focus:outline-none hover:bg-gray-200 hover:cursor-pointer rounded text-base font-semibold duration-300"
       >
         Throw Error
